getting player identity function
